+++ include/field_control_avatar.h
@@ -18,7 +18,7 @@ struct FieldInput
     bool8 fieldButtonToggleRun:1;
     bool8 input_field_1_5:1;
     bool8 input_field_1_6:1;
-    bool8 input_field_1_7:1;
+    bool8 fieldOpenDebugMenu:1;
     u8 dpadDirection;
 };