+++ include/strings/field_debug_menu.h
@@ -1,0 +1,41 @@
+#ifndef GUARD_LU_FIELD_DEBUG_MENU_STRINGS_H
+#define GUARD_LU_FIELD_DEBUG_MENU_STRINGS_H
+
+#include "gba/types.h"
+
+extern const u8 gText_lu_FieldDebugMenu_DisableTrainerLOS[];
+extern const u8 gText_lu_FieldDebugMenu_DisableWildEncounters[];
+extern const u8 gText_lu_FieldDebugMenu_FastTravel[];
+extern const u8 gText_lu_FieldDebugMenu_UseAnyBike[];
+extern const u8 gText_lu_FieldDebugMenu_UseAnyFieldMove[];
+extern const u8 gText_lu_FieldDebugMenu_UseAnyFishingRod[];
+extern const u8 gText_lu_FieldDebugMenu_SetTime[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather[];
+extern const u8 gText_lu_FieldDebugMenu_WalkThroughWalls[];
+
+extern const u8 gText_lu_FieldDebugMenu_UseAnyBike_Acro[];
+extern const u8 gText_lu_FieldDebugMenu_UseAnyBike_Mach[];
+
+extern const u8 gText_lu_FieldDebugMenu_UseAnyFishingRod_Old[];
+extern const u8 gText_lu_FieldDebugMenu_UseAnyFishingRod_Good[];
+extern const u8 gText_lu_FieldDebugMenu_UseAnyFishingRod_Super[];
+
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_SunnyCloudy[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Sunny[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Rain[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Snow[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Thunderstorm[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Fog[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Ashfall[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Sandstorm[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_FogDiagonal[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Underwater[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Overcast[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Drought[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Downpour[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_UnderwaterBubbles[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Abnormal[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Route119Cycle[];
+extern const u8 gText_lu_FieldDebugMenu_SetWeather_Route123Cycle[];
+
+#endif