+++ src/region_map.c
@@ -28,6 +28,8 @@
 #include "constants/rgb.h"
 #include "constants/weather.h"
 
+#include "lu/field_debug_menu.h"
+
 /*
  *  This file handles region maps generally, and the map used when selecting a fly destination.
  *  Specific features of other region map uses are handled elsewhere
@@ -1172,51 +1174,60 @@ static void RegionMap_InitializeStateBasedOnSSTidalLocation(void)
     sRegionMap->cursorPosY = gRegionMapEntries[sRegionMap->mapSecId].y + y + MAPCURSOR_Y_MIN;
 }
 
-static u8 GetMapsecType(u16 mapSecId)
-{
-    switch (mapSecId)
-    {
-    case MAPSEC_NONE:
-        return MAPSECTYPE_NONE;
-    case MAPSEC_LITTLEROOT_TOWN:
-        return FlagGet(FLAG_VISITED_LITTLEROOT_TOWN) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_OLDALE_TOWN:
-        return FlagGet(FLAG_VISITED_OLDALE_TOWN) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_DEWFORD_TOWN:
-        return FlagGet(FLAG_VISITED_DEWFORD_TOWN) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_LAVARIDGE_TOWN:
-        return FlagGet(FLAG_VISITED_LAVARIDGE_TOWN) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_FALLARBOR_TOWN:
-        return FlagGet(FLAG_VISITED_FALLARBOR_TOWN) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_VERDANTURF_TOWN:
-        return FlagGet(FLAG_VISITED_VERDANTURF_TOWN) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_PACIFIDLOG_TOWN:
-        return FlagGet(FLAG_VISITED_PACIFIDLOG_TOWN) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_PETALBURG_CITY:
-        return FlagGet(FLAG_VISITED_PETALBURG_CITY) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_SLATEPORT_CITY:
-        return FlagGet(FLAG_VISITED_SLATEPORT_CITY) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_MAUVILLE_CITY:
-        return FlagGet(FLAG_VISITED_MAUVILLE_CITY) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_RUSTBORO_CITY:
-        return FlagGet(FLAG_VISITED_RUSTBORO_CITY) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_FORTREE_CITY:
-        return FlagGet(FLAG_VISITED_FORTREE_CITY) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_LILYCOVE_CITY:
-        return FlagGet(FLAG_VISITED_LILYCOVE_CITY) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_MOSSDEEP_CITY:
-        return FlagGet(FLAG_VISITED_MOSSDEEP_CITY) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_SOOTOPOLIS_CITY:
-        return FlagGet(FLAG_VISITED_SOOTOPOLIS_CITY) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_EVER_GRANDE_CITY:
-        return FlagGet(FLAG_VISITED_EVER_GRANDE_CITY) ? MAPSECTYPE_CITY_CANFLY : MAPSECTYPE_CITY_CANTFLY;
-    case MAPSEC_BATTLE_FRONTIER:
-        return FlagGet(FLAG_LANDMARK_BATTLE_FRONTIER) ? MAPSECTYPE_BATTLE_FRONTIER : MAPSECTYPE_NONE;
-    case MAPSEC_SOUTHERN_ISLAND:
-        return FlagGet(FLAG_LANDMARK_SOUTHERN_ISLAND) ? MAPSECTYPE_ROUTE : MAPSECTYPE_NONE;
-    default:
-        return MAPSECTYPE_ROUTE;
-    }
+static u8 GetMapsecType(u16 mapSecId) {
+   if (mapSecId == MAPSEC_NONE) {
+      return MAPSECTYPE_NONE;
+   }
+   u16 flag     = 0;
+   u8  landmark = MAPSECTYPE_NONE;
+   switch (mapSecId) {
+      #define FLAG(name) \
+      case MAPSEC_##name: \
+         flag = FLAG_VISITED_##name; \
+         break;
+      
+      FLAG(LITTLEROOT_TOWN)
+      FLAG(OLDALE_TOWN)
+      FLAG(DEWFORD_TOWN)
+      FLAG(LAVARIDGE_TOWN)
+      FLAG(FALLARBOR_TOWN)
+      FLAG(VERDANTURF_TOWN)
+      FLAG(PACIFIDLOG_TOWN)
+      FLAG(PETALBURG_CITY)
+      FLAG(SLATEPORT_CITY)
+      FLAG(MAUVILLE_CITY)
+      FLAG(RUSTBORO_CITY)
+      FLAG(FORTREE_CITY)
+      FLAG(LILYCOVE_CITY)
+      FLAG(MOSSDEEP_CITY)
+      FLAG(SOOTOPOLIS_CITY)
+      FLAG(EVER_GRANDE_CITY)
+      
+      #undef FLAG
+      
+      case MAPSEC_BATTLE_FRONTIER:
+         flag     = FLAG_LANDMARK_BATTLE_FRONTIER;
+         landmark = MAPSECTYPE_BATTLE_FRONTIER;
+         break;
+      case MAPSEC_SOUTHERN_ISLAND:
+         flag     = FLAG_LANDMARK_SOUTHERN_ISLAND;
+         landmark = MAPSECTYPE_ROUTE;
+         break;
+      
+      default:
+         return MAPSECTYPE_ROUTE;
+   }
+   if (flag == 0) {
+      return MAPSECTYPE_ROUTE;
+   }
+   if (gFieldDebugMenuState.allow_fast_travel_anywhere || FlagGet(flag)) {
+      if (landmark != MAPSECTYPE_NONE)
+         return landmark;
+      return MAPSECTYPE_CITY_CANFLY;
+   }
+   if (landmark != MAPSECTYPE_NONE)
+      return MAPSECTYPE_NONE;
+   return MAPSECTYPE_CITY_CANTFLY;
 }
 
 u16 GetRegionMapSecIdAt(u16 x, u16 y)
@@ -1866,7 +1877,7 @@ static void CreateFlyDestIcons(void)
         {
             gSprites[spriteId].oam.shape = shape;
 
-            if (FlagGet(canFlyFlag))
+            if (gFieldDebugMenuState.allow_fast_travel_anywhere || FlagGet(canFlyFlag))
                 gSprites[spriteId].callback = SpriteCB_FlyDestIcon;
             else
                 shape += 3;
@@ -1892,7 +1903,7 @@ static void TryCreateRedOutlineFlyDestIcons(void)
 
     for (i = 0; sRedOutlineFlyDestinations[i][1] != MAPSEC_NONE; i++)
     {
-        if (FlagGet(sRedOutlineFlyDestinations[i][0]))
+        if (gFieldDebugMenuState.allow_fast_travel_anywhere || FlagGet(sRedOutlineFlyDestinations[i][0]))
         {
             mapSecId = sRedOutlineFlyDestinations[i][1];
             GetMapSecDimensions(mapSecId, &x, &y, &width, &height);