+++ src/strings/field_debug_menu.c
@@ -1,0 +1,36 @@
+#include "strings/field_debug_menu.h"
+
+const u8 gText_lu_FieldDebugMenu_DisableTrainerLOS[] = _("Disable trainer LOS");
+const u8 gText_lu_FieldDebugMenu_DisableWildEncounters[] = _("Disable wild encounters");
+const u8 gText_lu_FieldDebugMenu_FastTravel[] = _("Fast-travel");
+const u8 gText_lu_FieldDebugMenu_UseAnyBike[] = _("Use any bike");
+const u8 gText_lu_FieldDebugMenu_UseAnyFieldMove[] = _("Use any field move");
+const u8 gText_lu_FieldDebugMenu_UseAnyFishingRod[] = _("Use any fishing rod");
+const u8 gText_lu_FieldDebugMenu_SetTime[] = _("Set time of day");
+const u8 gText_lu_FieldDebugMenu_SetWeather[] = _("Set weather");
+const u8 gText_lu_FieldDebugMenu_WalkThroughWalls[] = _("Walk through walls");
+
+const u8 gText_lu_FieldDebugMenu_UseAnyBike_Acro[] = _("Acro Bike");
+const u8 gText_lu_FieldDebugMenu_UseAnyBike_Mach[] = _("Mach Bike");
+
+const u8 gText_lu_FieldDebugMenu_UseAnyFishingRod_Old[] = _("Old Rod");
+const u8 gText_lu_FieldDebugMenu_UseAnyFishingRod_Good[] = _("Good Rod");
+const u8 gText_lu_FieldDebugMenu_UseAnyFishingRod_Super[] = _("Super Rod");
+
+const u8 gText_lu_FieldDebugMenu_SetWeather_SunnyCloudy[] = _("Sunny + clouds");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Sunny[] = _("Sunny");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Rain[] = _("Rain");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Snow[] = _("Snow");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Thunderstorm[] = _("Thunderstorm");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Fog[] = _("Fog");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Ashfall[] = _("Ashfall");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Sandstorm[] = _("Sandstorm");
+const u8 gText_lu_FieldDebugMenu_SetWeather_FogDiagonal[] = _("Fog Diagonal");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Underwater[] = _("RS Underwater");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Overcast[] = _("Overcast");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Drought[] = _("RS Drought");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Downpour[] = _("RS Downpour");
+const u8 gText_lu_FieldDebugMenu_SetWeather_UnderwaterBubbles[] = _("Underwater");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Abnormal[] = _("Abnormal");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Route119Cycle[] = _("Route 119");
+const u8 gText_lu_FieldDebugMenu_SetWeather_Route123Cycle[] = _("Route 123");